reparation state controller

+++ MegaHerdt.Helpers/Helpers/ReparationClaimHelper.cs
@@ -13,32 +13,55 @@ namespace MegaHerdt.Helpers.Helpers
     public class ReparationClaimHelper
     {
         private readonly Repository<ReparationClaim> repository;
-        public ReparationClaimHelper(Repository<ReparationClaim> repository)
+        private readonly Repository<Reparation> repositoryReparation;
+        public ReparationClaimHelper(Repository<ReparationClaim> repository, Repository<Reparation> repositoryReparation)
         {
             this.repository = repository;
+            this.repositoryReparation = repositoryReparation;
         }
 
         public async Task<ReparationClaim> Create(ReparationClaim reparationClaim)
         {
-            return await this.repository.Add(reparationClaim);
+            if (validateReparationData(reparationClaim))
+            {
+                return await this.repository.Add(reparationClaim);
+            }
+            else { throw new Exception("reparation credentials are invalids"); }
+
         }
 
         public async Task Update(ReparationClaim reparationClaim)
         {
-            await this.repository.Update(reparationClaim);
+            if (validateReparationData(reparationClaim))
+            {
+                await this.repository.Update(reparationClaim);
+            }
+            else { throw new Exception("reparation credentials are invalids"); }
         }
 
         public async Task Delete(ReparationClaim reparationClaim)
         {
-            await this.repository.Delete(reparationClaim);
+            if (validateReparationData(reparationClaim))
+            {
+                await this.repository.Delete(reparationClaim);
+            }
         }
 
         public IQueryable<ReparationClaim> Get(Expression<Func<ReparationClaim, bool>> filter = null)
         {
             return repository.Get(filter)
                 .Include(x => x.Client)
-                .Include(x => x.Reparation);
+                .Include(x => x.Reparation)
+                .ThenInclude(x => x.Employee);
         }
 
+
+        private bool validateReparationData(ReparationClaim reparationClaim)
+        {
+            Expression<Func<Reparation, bool>> filter = x => x.Id == reparationClaim.ReparationId;
+            var reparation = repositoryReparation.Get(filter).FirstOrDefault();
+            var validClient = reparation != null && reparation.ClientId == reparationClaim.ClientId;
+            return validClient;
+        }
     }
 }
